+++ R/getSymbols.MarketWatch.R
@@ -39,8 +39,18 @@
     mw.from = format(as.Date(from),'%m/%d/%Y')
     mw.to = format(as.Date(to),'%m/%d/%Y')
     full_url <- paste0(endpoint,market,'/' ,Symbol.name,'/downloaddatapartial?startdate=',mw.from,'%2000:00:00&enddate=',mw.to,'%2023:59:59&frequency=p1d&csvdownload=true&countrycode=',countrycode)
-    df = fread(full_url)
-
+    if(verbose==T) print(full_url)
+    headers = add_headers('Accept' = 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
+                          'Accept-Language' = 'en-US,en;q=0.9,fr;q=0.8,ja;q=0.7,es;q=0.6',
+                          'Accept-Encoding' = 'gzip, deflate, br',
+                          'Connection' = 'keep-alive',
+                          'Host' = 'www.marketwatch.com',
+                          'Sec-Fetch-Dest' = 'document',
+                          'Sec-Fetch-Mode' = 'navigate',
+                          'Sec-Fetch-Site' = 'none',
+                          'User-Agent'= 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/17.4.1 Safari/605.1.15')
+    r <- GET(full_url,headers)
+    df = data.table(content(r))
     Symbol.name <-toupper(gsub('\\^','',Symbol.name))
     if(auto.assign)
       assign(Symbol.name,df,env)